WIP on a test framework based on script_tests.py
Copied script_tests.py and test files to a subdir.
Cut out test specfic code from original file and document.
