+++ t/py/convert_assyfile_to_contigset/test_files/AssemblyFile.json
